move list of nfts into separate component
